separate Class from main.dart
